pyproject.toml
[project]
name = "luceo"
version = "0.1.0"
description = "AI-powered addiction recovery support platform"
readme = "README.md"
license = { file = "LICENCE" }
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "anthropic>=0.52",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "cryptography>=44.0",
    "alembic>=1.14",
    "email-validator>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 8.0",
    "pytest-asyncio >= 0.24",
    "httpx >= 0.28",
    "aiosqlite >= 0.20",
    "ruff >= 0.9",
    "mypy >= 1.14",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
